refactor: removed enumerator itens types
The enumerator was removed to indicate if it is a folder or file, now this type of information will
come from outside the component. As well as the new type list property allowed to drop on each item
in the tree.

BREAKING CHANGE: Item type enumerator has been removed. Added a property to receive a list of
dropable items in each list item. Item type changed the type to any.